requirements.txt
aiohttp==3.7.3
aiopg==1.1.0
async-timeout==3.0.1
attrs==20.3.0
chardet==3.0.4
idna==2.10
multidict==5.1.0
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
six==1.15.0
SQLAlchemy==1.3.22
typing-extensions==3.7.4.3
yarl==1.6.3
